Reconstruct Queue.
Maintain a single min heap and add number of min heap when sufficient
number of elements greater than it's valence are in the queue.

This solution is not good in terms of time complexity. At each step we
need to check all the elements since this is maintained by a list.
We can do it using a linked list or even a simple dynamic list. Which
will decrease it's time complexity.