reconstructQueue based on the sorting order.
This one uses two keys and sorting and use a interesting trick to get
the output.

First the tallest person is added and then it is added based on the
index. All other elements are added based on height. but then u can
insert it based on lower index. it will go to the right position.