buildTree_inorder_postorder: using recursion
Main idea is to find boundaries for left and right subtree using
informatin from both inorder and preorder. This can give us full
information.

Use post order to find the parent. Search parent in inorder. Left of
parent in inorder is left subtree and right is right subtree. Use number
of nodes to find elements in postorder.... Do this recursively.