taskSchedular: Use greedy algorithm using two heap
We use two heaps to get the greedy solution. First heap selects the item
with highest frequency. But item can only exist in first heap if it can
be performed.
So when a item is performed, it's added to the other heap which will
keep the item until it's waiting peroid has expired. Both heaps are
modelled as a min heap by manipulating the values.