rotate matrix by 90 one liner python
Most important thing i learned was A=zip thingy reassigns and since we
are checking for inplace it fails and original array is not changed just
the A variable is assigned to new zip content. Hence inplace thing
fails.

otherwise it's quite simple * operator to pass each list to zip and
[::-1] to reverse the list.